pyproject.toml
[project]
name = "embodied-ai"
version = "0.1.0"
description = "Embodied AI PWA with camera-aware chat, autonomy, and optional TTS"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.0.0",
    "anthropic>=0.30.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.30.0",
    "httpx>=0.27.0",
    "mangum>=0.17.0",
]

[project.scripts]
embodied-ai = "embodied_ai.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/embodied_ai"]

[tool.hatch.build.targets.wheel.force-include]
"src/embodied_ai/web" = "embodied_ai/web"
